Conflicts:
	Steer/DigitizerWorkflow/CMakeLists.txt

+++ Detectors/MUON/MCH/Simulation/src/Response.cxx
@@ -105,7 +105,7 @@ double Response::chargefrac1d(float min, float max, double k2, double sqrtk3, do
   return 2. * k4 * (TMath::ATan(u2) - TMath::ATan(u1));
 }
 //______________________________________________________________________
-unsigned long Response::response(unsigned long adc)
+uint32_t Response::response(uint32_t adc)
 {
   //DecalibrateTrackerDigit functionality from
   //AliMuonDigitizerV3 in aliroot
